requirements.txt
Django>5.1
uritemplate>=3.0.0
pyyaml>=5.1
pygments>=2.7.0
markdown>=3.3.0
django-filter>1.0.1
django-guardian>=1.1.1
djangorestframework>=3.15.2
bcrypt>=4.2.0
psycopg2>=2.9.5
python-dotenv
django-extensions
djangorestframework-simplejwt
django-cors-headers
pillow